generate static open graph images during SSG
